+++ app/src/context/authorization/AuthorizationContextController.tsx
@@ -1,21 +1,28 @@
-import React from "react";
+import React, { useState } from "react";
 import { OAuthTokenStoreKey } from "api/oauth/oauth.types";
 import Cookies from "js-cookie";
 
 import { useToastContext } from "hooks/useToastContext/useToastContext";
 import { Typography } from "ui/typography/Typography";
 
-import { AuthorizationContextControllerProps } from "./AuthorizationContext.types";
+import { AccessTokens, AuthorizationContextControllerProps } from "./AuthorizationContext.types";
 import { AuthorizationContext } from "./AuthorizationContext";
 
 export const AuthorizationContextController = ({ children }: AuthorizationContextControllerProps) => {
+  const [accessTokens, setAccessTokens] = useState<AccessTokens>({});
+
   const toast = useToastContext();
 
   const verifyDropboxESignAuthorization = async () => {
     try {
       const dropboxESignAuthResponse = Cookies.get(OAuthTokenStoreKey.dropbox_esign);
 
-      JSON.parse(dropboxESignAuthResponse!);
+      const value = JSON.parse(dropboxESignAuthResponse!);
+
+      setAccessTokens((prev) => ({
+        ...prev,
+        [OAuthTokenStoreKey.dropbox_esign]: value.access_token,
+      }));
 
       toast.trigger({
         title: "Authorization successful",
@@ -30,6 +37,7 @@ export const AuthorizationContextController = ({ children }: AuthorizationContex
 
   const props = {
     verifyDropboxESignAuthorization,
+    accessTokens,
   };
 
   return <AuthorizationContext.Provider value={props}>{children}</AuthorizationContext.Provider>;
