+++ app/src/context/form/FormContext.types.ts
@@ -1,6 +1,8 @@
 import { FormApi } from "final-form";
 import { Dispatch, ReactNode, SetStateAction } from "react";
 
+import { ChatFormValues } from "app/chat/dropbox-chat/DropboxChat.types";
+
 export type FormState = FormApi<Record<string, any>, Partial<Record<string, any>>>;
 
 export type FormContextControllerProps = {
@@ -12,4 +14,5 @@ export type FormContextType = {
   setFieldValue: (field: string, text: string) => void;
   updateTextareaHeight: (id?: string) => { textarea: HTMLTextAreaElement; defaultHeight: string };
   resetTextareaHeight: () => void;
+  submit: (values: ChatFormValues) => Promise<void>;
 };
