+++ app/src/pages/api/chat/types.ts
@@ -1,5 +1,9 @@
 import { ChatContextMessage } from "context/message/MessageContext.types";
 
+export enum APIChatHeaderKeyNames {
+  x_dropbox_access_token = "x-dropbox-access-token",
+}
+
 export type DropboxESignRequest = {
   messages: ChatContextMessage[];
   currentMessage: ChatContextMessage;
