+++ app/src/providers/dropbox/dropbox.types.ts
@@ -1,0 +1,6 @@
+import * as DropboxSign from "@dropbox/sign";
+
+export type CreateEmbeddedSignatureRequestArgs = Pick<
+  DropboxSign.SignatureRequestCreateEmbeddedRequest,
+  "title" | "subject" | "message" | "signers"
+>;
